+++ catalyst/core/__init__.py
@@ -8,14 +8,7 @@
 # runner
 
 from .state import _State
-from .callback import (
-    Callback,
-    CallbackOrder,
-    CallbackNode,
-    MetricCallback,
-    MultiMetricCallback,
-    RaiseExceptionCallback,
-)
+from .callback import Callback, CallbackOrder, CallbackNode
 from .callbacks import *
 from .experiment import _Experiment
 from .runner import _Runner
