+++ catalyst/core/callbacks/__init__.py
@@ -2,7 +2,10 @@
 
 from .checkpoint import CheckpointCallback, IterationCheckpointCallback
 from .criterion import CriterionCallback
+from .early_stop import EarlyStoppingCallback
+from .exception import ExceptionCallback
 from .logging import ConsoleLogger, TensorboardLogger, VerboseLogger, MetricsManagerCallback
+from .metrics import MetricCallback, MultiMetricCallback, MetricAggregatorCallback
 from .optimizer import OptimizerCallback
 from .phase import PhaseManagerCallback
 from .scheduler import LRUpdater, SchedulerCallback
