+++ catalyst/dl/callbacks/__init__.py
@@ -12,6 +12,6 @@ from .metrics import (
     ClasswiseJaccardCallback, DiceCallback, F1ScoreCallback, IouCallback,
     JaccardCallback, MapKCallback, PrecisionRecallF1ScoreCallback
 )
-from .misc import ConfusionMatrixCallback, EarlyStoppingCallback
+from .confusion_matrix import ConfusionMatrixCallback
 from .mixup import MixupCallback
 from .scheduler import LRFinder
